+++ server/service/labels.go
@@ -3,6 +3,7 @@ package service
 import (
 	"context"
 
+	"github.com/fleetdm/fleet/v4/server/contexts/viewer"
 	"github.com/fleetdm/fleet/v4/server/fleet"
 )
 
@@ -159,6 +160,52 @@ func (svc *Service) GetLabel(ctx context.Context, id uint) (*fleet.Label, error)
 	}
 
 	return svc.ds.Label(ctx, id)
+}
+
+////////////////////////////////////////////////////////////////////////////////
+// List Labels
+////////////////////////////////////////////////////////////////////////////////
+
+type listLabelsRequest struct {
+	ListOptions fleet.ListOptions `url:"list_options"`
+}
+
+type listLabelsResponse struct {
+	Labels []labelResponse `json:"labels"`
+	Err    error           `json:"error,omitempty"`
+}
+
+func (r listLabelsResponse) error() error { return r.Err }
+
+func listLabelsEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	req := request.(*listLabelsRequest)
+	labels, err := svc.ListLabels(ctx, req.ListOptions)
+	if err != nil {
+		return listLabelsResponse{Err: err}, nil
+	}
+
+	resp := listLabelsResponse{}
+	for _, label := range labels {
+		labelResp, err := labelResponseForLabel(ctx, svc, label)
+		if err != nil {
+			return listLabelsResponse{Err: err}, nil
+		}
+		resp.Labels = append(resp.Labels, *labelResp)
+	}
+	return resp, nil
+}
+
+func (svc *Service) ListLabels(ctx context.Context, opt fleet.ListOptions) ([]*fleet.Label, error) {
+	if err := svc.authz.Authorize(ctx, &fleet.Label{}, fleet.ActionRead); err != nil {
+		return nil, err
+	}
+	vc, ok := viewer.FromContext(ctx)
+	if !ok {
+		return nil, fleet.ErrNoContext
+	}
+	filter := fleet.TeamFilter{User: vc.User, IncludeObserver: true}
+
+	return svc.ds.ListLabels(ctx, filter, opt)
 }
 
 func labelResponseForLabel(ctx context.Context, svc fleet.Service, label *fleet.Label) (*labelResponse, error) {
