+++ server/service/labels_test.go
@@ -8,6 +8,7 @@ import (
 	"github.com/fleetdm/fleet/v4/server/fleet"
 	"github.com/fleetdm/fleet/v4/server/test"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestLabelsWithDS(t *testing.T) {
@@ -18,6 +19,7 @@ func TestLabelsWithDS(t *testing.T) {
 		fn   func(t *testing.T, ds *mysql.Datastore)
 	}{
 		{"GetLabel", testLabelsGetLabel},
+		{"ListLabels", testLabelsListLabels},
 	}
 	for _, c := range cases {
 		t.Run(c.name, func(t *testing.T) {
@@ -41,4 +43,13 @@ func testLabelsGetLabel(t *testing.T, ds *mysql.Datastore) {
 	labelVerify, err := svc.GetLabel(test.UserContext(test.UserAdmin), label.ID)
 	assert.Nil(t, err)
 	assert.Equal(t, label.ID, labelVerify.ID)
+}
+
+func testLabelsListLabels(t *testing.T, ds *mysql.Datastore) {
+	svc := newTestService(ds, nil, nil)
+	require.NoError(t, ds.MigrateData(context.Background()))
+
+	labels, err := svc.ListLabels(test.UserContext(test.UserAdmin), fleet.ListOptions{Page: 0, PerPage: 1000})
+	require.NoError(t, err)
+	require.Len(t, labels, 7)
 }
